# :sparkles: Add customTags in logger middleware Config