@Skyenought 2022-11-01
> Can you send here old and new benchmark comparision

Seems to have slowed down :-(

old
```
goos: linux
goarch: amd64                                     
pkg: github.com/gofiber/fiber/v2/middleware/logger
cpu: Intel(R) Core(TM) i5-1035G1 CPU @ 1.00GHz
Benchmark_Logger
Benchmark_Logger-2       5582337               219.6 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       5000182               237.4 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       5331586               215.4 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       5741889               209.6 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v2/middleware/logger   5.683s

PASS
```
new
```
goos: linux  
goarch: amd64
pkg: github.com/gofiber/fiber/v2/middleware/logger
cpu: Intel(R) Core(TM) i5-1035G1 CPU @ 1.00GHz
Benchmark_Logger
Benchmark_Logger-2       1649736               706.7 ns/op           320 B/op          7 allocs/op
Benchmark_Logger-2       1761847               732.1 ns/op           320 B/op          7 allocs/op
Benchmark_Logger-2       1704952               673.5 ns/op           320 B/op          7 allocs/op
Benchmark_Logger-2       1722056               674.2 ns/op           320 B/op          7 allocs/op
PASS
ok      github.com/gofiber/fiber/v2/middleware/logger   7.636s

PASS
```