@Skyenought 2022-11-02
> Ok then just try to improve a little bit



> Ok then just try to improve a little bit

It's hard to optimize this code at my level, so it looks like I'll have to close this PR :-<