@Skyenought 2022-11-04
use switch
```sh
goos: linux
goarch: amd64                                     
pkg: github.com/gofiber/fiber/v2/middleware/logger
cpu: Intel(R) Core(TM) i5-1035G1 CPU @ 1.00GHz
Benchmark_Logger
Benchmark_Logger-2       5582337               219.6 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       5000182               237.4 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       5331586               215.4 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       5741889               209.6 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v2/middleware/logger   5.683s

PASS
```

use map
```sh
goos: linux                                       
goarch: amd64                                     
pkg: github.com/gofiber/fiber/v2/middleware/logger
cpu: Intel(R) Core(TM) i5-1035G1 CPU @ 1.00GHz    
Benchmark_Logger
Benchmark_Logger-2       4919536               244.3 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       4852894               256.5 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       4817518               243.4 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       4551956               249.3 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v2/middleware/logger   5.968s

```