@Skyenought 2022-11-10
> No, will look at the pull request in the next few days Currently there is a lot to do

I'm sorry to bother you, but I'm just confused about the PR not passing.