@Skyenought 2022-11-15
Now logger support custom tags like "reqHeader:x-key"

```
goos: linux                                       
goarch: amd64                                     
pkg: github.com/gofiber/fiber/v2/middleware/logger
cpu: Intel(R) Core(TM) i5-1035G1 CPU @ 1.00GHz
Benchmark_Logger
Benchmark_Logger-2       4537462               266.2 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       4529635               278.0 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       4531717               267.3 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       4247362               276.5 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v2/middleware/logger   5.962s
```