@Skyenought 2022-11-17
before change:
```sh
goos: linux
goarch: amd64
pkg: github.com/gofiber/fiber/v2/middleware/logger
cpu: Intel(R) Core(TM) i5-1035G1 CPU @ 1.00GHz
Benchmark_Logger
Benchmark_Logger-2       3122877               386.0 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       3019996               385.9 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       3146631               382.3 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       3159733               388.7 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v2/middleware/logger   6.380s
```

after change:
```
goos: linux  
goarch: amd64
pkg: github.com/gofiber/fiber/v2/middleware/logger
cpu: Intel(R) Core(TM) i5-1035G1 CPU @ 1.00GHz
Benchmark_Logger
Benchmark_Logger-2       3049252               373.8 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       3194596               406.7 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       3253861               368.0 ns/op             0 B/op          0 allocs/op
Benchmark_Logger-2       3207115               371.3 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/gofiber/fiber/v2/middleware/logger   6.375s
```
Seems to have improved a little. @ReneWerner87