@ReneWerner87 2022-11-18
After the usage of syncPool and data structs

benchmark with format `"${bytesReceived} ${bytesSent} ${status} ${reqHeader:test}"` and local in handler:
master:
```go
Benchmark_Logger-12    	 4105185	       293.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4130638	       290.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4085211	       288.9 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4149218	       292.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4114450	       290.6 ns/op	       0 B/op	       0 allocs/op
```
feature branch:
```go
Benchmark_Logger-12    	 4293021	       290.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4115019	       287.7 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4231435	       287.5 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4059115	       284.3 ns/op	       0 B/op	       0 allocs/op
Benchmark_Logger-12    	 4272568	       281.6 ns/op	       0 B/op	       0 allocs/op
```
---
benchmark with default format `"[${time}] ${status} - ${latency} ${method} ${path}\n"` and local in handler:
master:
```go
Benchmark_Logger-12    	 1840008	       628.7 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1882804	       636.4 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1841458	       684.7 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1883836	       662.0 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1851812	       643.3 ns/op	       8 B/op	       1 allocs/op
```
feature branch:
```go
Benchmark_Logger-12    	 1907866	       663.8 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1835814	       641.3 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1710475	       633.3 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1789392	       652.6 ns/op	       8 B/op	       1 allocs/op
Benchmark_Logger-12    	 1888110	       645.6 ns/op	       8 B/op	       1 allocs/op
```